+++ frontend/package.json
@@ -47,9 +47,9 @@
     "react-router": "^6.0.0-beta.0",
     "react-router-dom": "^6.0.0-beta.0",
     "react-scripts": "^4.0.0",
+    "react-swipeable-views": "^0.13.9",
     "uuid": "^8.3.0",
     "yup": "^0.29.3"
   },
-  "devDependencies": {
-  }
+  "devDependencies": {}
 }
